feat: upgrade Mangle v0.4.0 → HEAD with stratified eval and native float/time/duration types
- Upgrade mangle dependency to v0.4.1-0.20260211131615-299701686e7a
- Replace EvalProgramWithStats with EvalStratifiedProgramWithStats across
  kernel_eval.go, engine.go, differential.go, and test helpers
- Cache strata/predToStratum in RealKernel and Engine structs
- Use native ast.Float64(), ast.Time(), ast.Duration() type constructors
  in types.go, engine.go, and virtual_store_graph.go
- Fix 4 production sites passing float64 to integer-schema predicates:
  action_validator.go (Confidence 0.0-1.0 → 0-100 int),
  virtual_store_predicates.go (avgDuration → int64),
  virtual_store_workflows.go (0.25 → int64(25)),
  perception/taxonomy.go (remove unnecessary float64 cast)
- Fix 4 ignored json.Unmarshal errors in lsp.go
- Remove debug logging from virtual_store_test.go
- Use upstream WithCreatedFactLimit gas limiting

+++ internal/core/virtual_store_predicates.go
@@ -364,7 +364,7 @@ func (v *VirtualStore) QueryTraceStats(shardType string) ([]Fact, error) {
 				shardType,
 				successCount,
 				failCount,
-				avgDuration,
+				int64(avgDuration), // Milliseconds as integer per schema
 			},
 		},
 	}
